Added "next" as a dynamic command
